Using a mock for CountryCodeSignalProvider (instead of interface with real and test implementations).
