--- app/src/org/commcare/utils/DefaultCountryCodeSignalProvider.kt
@@ -1,25 +1,0 @@
-package org.commcare.utils
-
-import android.content.Context
-import android.telephony.TelephonyManager
-import androidx.core.os.ConfigurationCompat
-import org.commcare.CommCareApplication
-
-/**
- * Production implementation that reads country signals from Android system services.
- */
-class DefaultCountryCodeSignalProvider() : CountryCodeSignalProvider {
-    private val appContext: Context = CommCareApplication.instance()
-    private val telephonyManager: TelephonyManager? =
-        appContext.getSystemService(Context.TELEPHONY_SERVICE) as? TelephonyManager
-
-    override val simCountryIso: String
-        get() = telephonyManager?.simCountryIso ?: ""
-
-    override val networkCountryIso: String
-        get() = telephonyManager?.networkCountryIso ?: ""
-
-    override val localeCountry: String
-        get() = ConfigurationCompat.getLocales(appContext.resources.configuration)
-            .get(0)?.country ?: ""
-}
